Factor out SortingTableHead
